refactor avatar utils
